+++ parser.uni/jyf-parser.vue
@@ -67,7 +67,7 @@
 	 * @event {Function} imgtap 图片点击事件
 	 * @event {Function} linkpress 链接点击事件
 	 * @author JinYufeng
-	 * @version 20201014
+	 * @version 20201029
 	 * @listens MIT
 	 */
 	export default {